+++ dds/DCPS/security/UtilityImpl.cpp
@@ -13,7 +13,8 @@ UtilityImpl::~UtilityImpl() {}
 
 void UtilityImpl::generate_random_bytes(void* ptr, size_t size)
 {
-  int rc = RAND_bytes(static_cast<unsigned char*>(ptr), size);
+  int rc = RAND_bytes(static_cast<unsigned char*>(ptr),
+                      static_cast<int>(size));
 
   if (rc != 1) {
     unsigned long err = ERR_get_error();
@@ -27,7 +28,10 @@ void UtilityImpl::generate_random_bytes(void* ptr, size_t size)
 
 void UtilityImpl::hmac(void* out, void const* in, size_t size, const std::string& password) const
 {
-  unsigned char* digest = HMAC(EVP_sha1(), password.c_str(), password.size(), static_cast<const unsigned char*>(in),size, NULL, NULL);
+  unsigned char* digest = HMAC(EVP_sha1(), password.c_str(),
+                               static_cast<int>(password.size()),
+                               static_cast<const unsigned char*>(in),
+                               static_cast<int>(size), NULL, NULL);
   memcpy(out, digest, 20);
 }
